+++ packages/council-deploy/src/deployments/mainnet.deployments.json
@@ -1,0 +1,4 @@
+{
+  "chainId": 1,
+  "deployments": []
+}
