+++ packages/council-deploy/src/vaults/lockingVault/promptChooseLockingVaultAndDepositTokens.ts
@@ -3,7 +3,8 @@ import { Wallet } from "ethers";
 import { formatUnits, parseUnits } from "ethers/lib/utils";
 import hre from "hardhat";
 import prompt from "prompt";
-import { getDeploymentForContract, goerliDeployments } from "src/deployments";
+import { goerliDeployments } from "src/deployments";
+import { getDeploymentForContract } from "src/deployments/getDeployments";
 import { promptString } from "src/base/prompts/promptString";
 import { promptMint } from "src/votingToken/promptMint";
 import { promptNumber } from "src/base/prompts/promptNumber";
