+++ packages/council-deploy/src/votingToken/promptMintFromListOfVotingTokens.ts
@@ -3,7 +3,8 @@ import { Wallet } from "ethers";
 import hre from "hardhat";
 import prompt from "prompt";
 import { promptString } from "src/base/prompts/promptString";
-import { getDeploymentForContract, goerliDeployments } from "src/deployments";
+import { goerliDeployments } from "src/deployments";
+import { getDeploymentForContract } from "src/deployments/getDeployments";
 import { DeploymentInfo } from "src/deployments/types";
 import { getVotingTokens } from "src/votingToken/getVotingTokens";
 import { promptMint } from "src/votingToken/promptMint";
